Change mayor & minister parsing and storing
