+++ src/main/java/de/hysky/skyblocker/utils/Utils.java
@@ -6,7 +6,6 @@ import com.mojang.util.UndashedUuid;
 import de.hysky.skyblocker.events.SkyblockEvents;
 import de.hysky.skyblocker.mixins.accessors.MessageHandlerAccessor;
 import de.hysky.skyblocker.skyblock.item.MuseumItemCache;
-import de.hysky.skyblocker.utils.scheduler.Scheduler;
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.azureaaron.hmapi.data.server.Environment;
@@ -27,14 +26,12 @@ import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Util;
-import org.apache.http.client.HttpResponseException;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.time.Instant;
 import java.util.Collections;
-import java.util.concurrent.CompletableFuture;
 
 /**
  * Utility variables and methods for retrieving Skyblock related information.
@@ -77,10 +74,6 @@ public class Utils {
     private static String locationRaw = "";
     @NotNull
     private static String map = "";
-    private static boolean mayorTickScheduled = false;
-    private static int mayorTickRetryAttempts = 0;
-    private static String mayor = "";
-    private static String minister = "";
 
     /**
      * @implNote The parent text will always be empty, the actual text content is inside the text's siblings.
@@ -193,30 +186,7 @@ public class Utils {
         return map;
     }
 
-    /**
-     * @return the current mayor as cached on skyblock join.
-     */
-    @NotNull
-    public static String getMayor() {
-        return mayor;
-    }
-
-    /**
-     * @return the current minister as cached on skyblock join.
-     */
-    @NotNull
-    public static String getMinister() {
-        return minister;
-    }
-
     public static void init() {
-        SkyblockEvents.JOIN.register(() -> {
-            if (!mayorTickScheduled) {
-                tickMayorCache();
-                scheduleMayorTick();
-                mayorTickScheduled = true;
-            }
-        });
         ClientReceiveMessageEvents.ALLOW_GAME.register(Utils::onChatMessage);
         ClientReceiveMessageEvents.GAME_CANCELED.register(Utils::onChatMessage); // Somehow this works even though onChatMessage returns a boolean
         ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> onDisconnect());
@@ -503,45 +473,6 @@ public class Utils {
         }
 
         return true;
-    }
-
-    private static void scheduleMayorTick() {
-        long currentYearMillis = SkyblockTime.getSkyblockMillis() % 446400000L; //446400000ms is 1 year, 105600000ms is the amount of time from early spring 1st to late spring 27th
-        // If current time is past late spring 27th, the next mayor change is at next year's spring 27th, otherwise it's at this year's spring 27th
-        long millisUntilNextMayorChange = currentYearMillis > 105600000L ? 446400000L - currentYearMillis + 105600000L : 105600000L - currentYearMillis;
-        Scheduler.INSTANCE.schedule(Utils::tickMayorCache, (int) (millisUntilNextMayorChange / 50) + 5 * 60 * 20); // 5 extra minutes to allow the cache to expire. This is a simpler than checking age and subtracting from max age and rescheduling again.
-    }
-
-    private static void tickMayorCache() {
-        CompletableFuture.supplyAsync(() -> {
-            try {
-                Http.ApiResponse response = Http.sendCacheableGetRequest("https://api.hypixel.net/v2/resources/skyblock/election", null); //Authentication is not required for this endpoint
-                if (!response.ok()) throw new HttpResponseException(response.statusCode(), response.content());
-                JsonObject json = JsonParser.parseString(response.content()).getAsJsonObject();
-                if (!json.get("success").getAsBoolean()) throw new RuntimeException("Request failed!"); //Can't find a more appropriate exception to throw here.
-                return json.get("mayor").getAsJsonObject();
-            } catch (Exception e) {
-                throw new RuntimeException(e); //Wrap the exception to be handled by the exceptionally block
-            }
-        }).exceptionally(throwable -> {
-            LOGGER.error("[Skyblocker] Failed to get mayor status!", throwable.getCause());
-            if (mayorTickRetryAttempts < 5) {
-                int minutes = 5 << mayorTickRetryAttempts; //5, 10, 20, 40, 80 minutes
-                mayorTickRetryAttempts++;
-                LOGGER.warn("[Skyblocker] Retrying in {} minutes.", minutes);
-                Scheduler.INSTANCE.schedule(Utils::tickMayorCache, minutes * 60 * 20);
-            } else {
-                LOGGER.warn("[Skyblocker] Failed to get mayor status after 5 retries! Stopping further retries until next reboot.");
-            }
-            return new JsonObject(); //Have to return a value for the thenAccept block.
-        }).thenAccept(result -> {
-            if (!result.isEmpty()) {
-                mayor = result.get("name").getAsString();
-                minister = result.getAsJsonObject("minister").get("name").getAsString();
-                LOGGER.info("[Skyblocker] Mayor set to {}, minister set to {}.", mayor, minister);
-                scheduleMayorTick(); //Ends up as a cyclic task with finer control over scheduled time
-            }
-        });
     }
 
     /**
